update to Stripe

+++ src/main/java/com/turing/ecommerce/DTO/StripePayObject.java
@@ -19,7 +19,7 @@ public class StripePayObject {
 	
 	
 	private String status;
-	private Integer amount;
+	private Long amount;
 	private String description;
 	private Integer orderId;
